Updated contributions after coercions in modules are imported only if the module is explicitly imported (r13310). The transition is rather smooth except for Algebra and LinAlg (which depends on Algebra).
For Algebra, because "require"s are not done in the order as "import"s (*),
the ambiguous paths are resolved quite differently, leading to
definitions not resolved using to the same coercion paths and later
calls to "auto" failing. The only solution I found was to add the
option "Set Automatic Coercions Import".

(*) The "require"s are done in the order the recursive dependencies
are first known while the "import"s are done in the order the
recursive dependencies are last known. Otherwise said when a require
occurs several time in the expansion of the dependencies graph, the
first one is kept and the redundant occurences are skipped. But when
an import occurs several time in the expansion of the dependencies
graph, only the last one is kept and the redundant occurences coming
first are skipped (the latter is an optimization to avoid doing too
much imports when only the last one should be visible in theory; in
practise, not all imports are commutative: in particular, adding
ambiguous coercion paths is not commutative).


Update wrt changes in notations and binders
Ssreflect: change in type Topconstr.interpretation, renaming of
  binder_let into closed_binder, locations kept in binders
Interface: change in arguments of CNotation, locations kept in
  binders + (formal) completion of pattern-matching on vernac_expr


Adapt to new [consider_remaining_unif_problems] interface.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@2364 305884b8-be16-0410-aa53-a3be2363cef1