Previous commit r2373, aka: https://gforge.inria.fr/scm/viewvc.php?view=rev&root=coqfinitgroup&revision=2373
was coq-contribs-commits:
    Fixed bug of Prenex Implicits (wrong ExplByName flags)

    Patch from Assia.

    git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq-contribs/branches/v8.3/Saclay/Ssreflect@1270 0cf17b13-060f-0410-b1b1-c666bec9822a

This commit:

Update with respect to the change of interface of impargs.ml
consecutive to addition of multiple implicit arguments sequences (r13492).
Note: nothing done to support multiple sequences in Ssreflect and
Interface except correct compilation of the ml files.


 Author:    herbelin <herbelin@0cf17b13-060f-0410-b1b1-c666bec9822a>

Staging area trunk/ssreflect/ssreflect12_v8.3 should now reflect the
state of the coq-contribs repository (this is not the case for
ssreflect/ssreflect12_trunk)

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@2461 305884b8-be16-0410-aa53-a3be2363cef1