pf_abs_evars_pirrel calls ssrautoprop
ssrautoprop is a tactic defined in Ltac, that defaults to trivial
but can be overridden, that is called to close evars in Prop that
occurr in other goals. E.g. if (p : P 1) is in the proof context,
and the goals to be generated are (P 1) and (forall Hyp : P 1, Q Hyp),
the system only generates only (Q p), assumoin trivial finds p as
proof for (P 1).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@2553 305884b8-be16-0410-aa53-a3be2363cef1