Add generalization support to congr.
The following syntax is now accepted: congr (_ == 0): H.
On a goal G, the result is equivalent to the following line:

  move: H; cut (H = G); [ move=> ->; by apply | congr (_ == 0) ].

As a bonus, the type of the equality on which congr is called is
used to give a type to the congruence argument, that means that
coercions like is_true should be inserted automatically. For example
( _ == 0) is a good congruence context for

  is_true (x == 0) = is_true (y == 0)

since the top = is on Prop, and "morally" type inference runs on

  (_ == 0 : Prop)

so is_true is inserted for you.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@2608 305884b8-be16-0410-aa53-a3be2363cef1