Adding lemma seperableFadjoinExtend.
This commit reverses the previous commit.
This commit relaxes the definition of Derivation droping the (D @: K <= K) requirement.
That allows us to prove the lemma subvDerivation.
The rest of the proofs still goes through with easy modifications (mostly droping parts of the proof).
This commit replaces ((Fadjoin K x) y) with ((Fadjion K y) x).  While these
two spaces are equal, the later is actually a more convienent for given that
y is the one that is seperable over K.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@2954 305884b8-be16-0410-aa53-a3be2363cef1