smooth transition to elim interpreting last dgen (the one after /)
ssrelim accepts the "what" argument in two formats: `EConstr and `EGen.
`EConstr passes an already interpreted constr, used only by
simplest_case/elim and by "case/v: x" that passes "(v x)".

Passing the pattern after the "/" in a dependent elimination as a ssrgen
makes it easier to handle it as any other pattern before the "/".

Nevertheless, if the eliminator is omitted, one has to process the last
ssrgen to figure out the inductive type and consequently the eliminator,
thus part of the advantage is gone and the code is as messy as before.

Since we now process the last ssrgen manually, things like
"elim/big_ind2: {-}_" do work, even if the pattern _ is indetermined,
but can be inferred and matches the goal (but is not abstracted thanks
to {-} and so, the other argument of the predicate is matched against an
unaltered goal).

Test file elim2.v covers some of the odd cases reported by Georges, even
is I was unable to reproduce, and thus test, all of them.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@3167 305884b8-be16-0410-aa53-a3be2363cef1