"set" and ":" support rewrite patterns with explicit redex:
   t
   X in t
   e in X in t
   e as X in t

Pattern interpretation recognizes the special, illtyped, term
(_ : fun X => t) as "X in t" so that we can define notations for
standard locations, for example RHS for "X in _ = X".

Patterns must be parenthesized. Notations for patterns with explicit
redex are reserved in ssreflect.v as well as notations for LHS and RHS:

  Reserved Notation "( a 'in' b )"        (at level 0).
  Reserved Notation "( a 'as' b )"        (at level 0).
  Reserved Notation "( a 'in' b 'in' c )" (at level 0).
  Reserved Notation "( a 'as' b 'in' c )" (at level 0).
  Notation "'RHS'" := (_ : fun X => _ = X).
  Notation "'LHS'" := (_ : fun X => X = _).

In the context of "set" and ":" pattern are preferred over terms. If
the user defines an interpretation for, say, "( _ in _ in _ )" he has 2
ways to get it interpreted as such: putting extra paretheses around it,
or specify another scope.

Examples:
  set t := (X in t).
  set t := (t in X in t).
  set t := (t as X in t).
  set t := {2}(t in RHS).
  set t := RHS.
  move: {x} (x.+1) {1}x y (x.+1 in RHS).
  move E: ((x.+1 in y)) => w.
  move E: (x.+1 in y)%myscope => w.

Todo: elim does not allow for these patterns (yet)

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@3216 305884b8-be16-0410-aa53-a3be2363cef1