Getting rid of spurious calls to the inj tactic
The intro pattern [], when used on on equalities (comprising boolean
propositions like (is_true p)), calls the inj tactic.
That tactic was not failing but behaving as generalize
(thus a no-op in the intro pattern context) in case there was
nothing to inject (like in y = true). Now the inj tactic fails
in these cases thus these spurious [] can be detected and the
library has been cleaned up. Only Coq-trunk detects these errors.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@3448 305884b8-be16-0410-aa53-a3be2363cef1