Synchro commit for Cyril: A (dirty) proof of quantifier elimination reduction for ordered integral domains.
Up to four admitted lemmas one of which being obviously false (see comment). 
To be fixed tomorrow.

Also the reduction to Lt is not that convenient (we use Gt in the
paper). Feel free to change that.

The file is intentionally not in Make since it is not clean enough in
its current state.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@3456 305884b8-be16-0410-aa53-a3be2363cef1