Flexible pattern matching fixed
Disclaimer: the code is really intricate, and the choice
of variable name does not really help to understand their
meaning. Moreover it is a genuine example of spaghetti code,
so I may be wrong. Assuming I'm right, the following holds:

The filter_upat function was setting the i0 variable
to the wrong value. In particular i0 was set to the number of
_pattern_ arguments and not the number of the arguments
of the application that is being matched against the pattern
(that by the way is already known to the caller).
To make it short a pattern [_ a] versus a term (f t a) was
never successfully HO matched since only f, and not (f t),
was aligned and unified with _. In practice the loop on the
!i0 value was always executed once, even if the pattern head
was flexible.

This bug has been around since the very beginning (well, it was
introduced together with this function in 2009-06-26), but was not
present in filter_upat_FO. Thus all patterns with a flexible head used
to work only when the FO phase matches (i.e. always on simple types) or
when the head _ needs to be unified with exactly one symbol.

Let's see if I broke something with this fix...

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@3574 305884b8-be16-0410-aa53-a3be2363cef1