Minimal changes for the new evar unfolding strategy
Most of the problems are due to the fact that the new
algorithm unfolds less often, so extra /foo here and there.
In a couple of places ?foo was diverging, didn't check why,
but the new unif alg is stronger, so it could happen.
The big question is why in integral_char a congr now
takes forever, while the semantically no-op code I've
added around it makes it super fast. Georges?

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@3757 305884b8-be16-0410-aa53-a3be2363cef1