Fix "rewrite [in foo]bar" anomaly when no foo in Goal.
Tested on v8.4 only... fingers crossed.

This made a bug in "unlock" show up in generic quotient, that used to
unlock something that was not locked. Now this is an error, so the .v
file was fixed as well. Also, rewriting with x=x was not always considered
an error (progress check) because an anomaly was raised instead of NoMatch.
I guess the anomaly was caught by a catch all _ somewhere...

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@3894 305884b8-be16-0410-aa53-a3be2363cef1