now complex form an algebraically closed field, we remove the axiom in algC, c'est cadeau !
The proof is using only linear algebra, and is inspired from a papier by
Harm Derksen. This comes with a brief formalization of *skew matrices*
and *companion matrices* (in the file complex.v for now). A
formalization of hermitian matricies could slightly simplify the proof
and may speed up the compilation time (see Lemma5).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@4041 305884b8-be16-0410-aa53-a3be2363cef1