patterns with (X in t) part fix
clear unnecessary occurrences of X in the context of
evars in t to work around 8.4/trunk habit of potentially using
the value of X to resolve these evars.

X just denotes a set of holes, not a piece of context. It is encoded
as a let in, hence extending the context of the evars in the pattern,
to ease the interpretation of the pattern and the recognition of the
named hole(s) the user specified.

since 8.4, some evars are not assigned immediately, but later.
When this happens, the context (the list of candidates) is traversed
in reverse order, so X is preferred to, say, "a" if the pattern matched
a term that has "a" where X occurs. Note that this patch is not 100%
complete since, in the following example, typing prevents X to be cleared
from the context of the meta that matches (t x), hence this x is grabbed
too:

Require Import ssreflect ssrfun ssrbool.

Axiom T : nat -> Type.
Axiom t : forall n, T n.
Axiom f : forall n, T n -> nat.

Lemma foo (x : nat) : x = f x (t x) + x.
set w := (X in _ = f _ (_ : T X) + _).

-->

w := x
========
x = f w (t w) + x


It may be considered as a feature in this case ;-)

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@4155 305884b8-be16-0410-aa53-a3be2363cef1