Respect eta expansions in pattern matching.
Imagine the goal is

  P [eta f] = 0

The tactic

  set F := (F in P [eta F])

now changes the goal into

  F := f
  =====
  P [eta F] = 0

Since 8.4 introduced eta, f and [eta f] are considered the same by
the conversion algorithm, hence once the pattern instantiation phase
instantiates the F with f, it finds an occurrence of F at [eta f] too,
resulting in

  F := f
  =====
  P F = 0

A special case has been added so that eta expansion is
skipped trough by the occurrences finding procedure. More precisely
when [eta t'] is compared with t, if t and t' are convertible the
eta expansion is skipped trough.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@4175 305884b8-be16-0410-aa53-a3be2363cef1