bigop redefined to better support notations (only 1 binder)
  \big[op/idx]_(i <- r | P) F   ===   bigop idx op r (fun i => BigBody P F)

This comes with BIG_P and BIG_F as predefined patterns, to be used like
that:

  set F := BIG_F

or the more precise (if there are many \big around)

  set F := (fun i => m * _ as BIG_F)

to grab from a goal like:

  \sum_(n <= i < m | odd i) m * (e + i)

the funcion/predicate and obtain:

  P := fun i => odd i
  F := fun i => m * (e + i)
  =============================
  \sum_(m <= i < m | P i) F i

Naming F or P does not destroy the notation, since the binder is still
there. BigBody is an inductive constructor eaten by reducebig, so unless
bigop is unlocked it is there to stay. Unlocking bigop and /= makes it
vanish (explicit iota redex).

The only annoyance is that type inference is weak, and to infer a type
for F knowing its expected type (that is given by op or idx) one has to
pass a dummy argument to BigBody to force a type unification before
processing the last argument (that is F).

Matita would push the typing constraint from the context since BigBody
is an inductive constructor with one parameter that is known to the
context. Ex:

   Inductive bigbody R := BigBody of bool & R.
   Check BigBody. (* forall R : Type, bool -> R -> bigbody R *)

   refine (fun i => BigBody ?R P F  :  I -> bigbody R).

?R could be assigned to R before touching F that has expected type ?R.
This cound even be user customizable as once Georges suggested.

In MathComp this kind of inference is important because knowing the
expected type helps inserting coercions, like the one from bool to nat
or the one from groups to sets. Hence BigBody takes an extra argument,
that notations set to the op, to force the assignement of ?R before
processing F. See bigop.v and relative comment.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@4179 305884b8-be16-0410-aa53-a3be2363cef1