fix apply regression w.r.t. ssr 1.4 (symptom: more goals opened)
This is due to a known change in the refiner of 8.4, and the fact that
in ssr 1.5 (to avoid etas expansions in the proof terms, now visible if
one does have@) I had to plug to this new refiner. In 8.5 the tactic
shelve_unifiables hides goals occurring in other goals (that can be
solved by unifying a goal (type) with something. It is used to preserve
backward compatibility. The code was ready to call this tactic after
apply in ssr/trunk/, but there is no such code in 8.4, and I was too
lazy to back port it.  Now the plugin for 8.4 and trunk behave the same
and behave as in ssr 1.4.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/coqfinitgroup/trunk@4773 305884b8-be16-0410-aa53-a3be2363cef1